+++ packages/apollo-link-http/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change log
 
+### vNext
+
+- If `name` or `version` client awareness settings are found in the
+  incoming `operation` `context`, they'll be extracted and added as headers
+  to all outgoing requests. The header names used (`apollographql-client-name`
+  and `apollographql-client-version`) line up with the associated Apollo Server
+  changes made in https://github.com/apollographql/apollo-server/pull/1960.  <br/>
+  [@hwillson](http://github.com/hwillson) in [#872](https://github.com/apollographql/apollo-link/pull/872)
+
 ### 1.5.5
 - Added `graphql` 14 to peer and dev deps; Updated `@types/graphql` to 14  <br/>
   [@hwillson](http://github.com/hwillson) in [#789](https://github.com/apollographql/apollo-link/pull/789)